fixing typo and Id data-type

+++ Repositories/Repository/Repository.cs
@@ -17,7 +17,7 @@ namespace Event.Repositories.Repository
             Context = context;
         }
 
-        public TEntity Get(int id)
+        public TEntity Get(long id)
         {
             // Here we are working with a DbContext, not SurveyContext. So we don't have DbSets 
             // such as Courses or Authors, and we need to use the generic Set() method to access them.
